+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Allow for skipping an ongoing DMA transfer if not using double buffering.
 - Change DMA traits to `embedded-dma`.
 - Use bitbanding during clock enabling and peripheral reset to avoid data races.
+- Add missing `Write` implementation for `Serial` and implemented better error handling.
 
 ### Added